+++ lib/hubspot/codegen/cms/source_code/api/content_api.rb
@@ -1,7 +1,7 @@
 =begin
 #CMS Source Code
 
-#Endpoints for interacting with files in the CMS Developer File System. These files include HTML templates, CSS, JS, modules, and other assets which are used to create CMS content.
+#No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 
 The version of the OpenAPI document: v3
 
@@ -100,7 +100,7 @@ module Hubspot
         # @param environment [String] The environment of the file (\&quot;draft\&quot; or \&quot;published\&quot;).
         # @param path [String] The file system location of the file.
         # @param [Hash] opts the optional parameters
-        # @option opts [File] :file The file to upload.
+        # @option opts [File] :file 
         # @return [AssetFileMetadata]
         def create(environment, path, opts = {})
           data, _status_code, _headers = create_with_http_info(environment, path, opts)
@@ -112,7 +112,7 @@ module Hubspot
         # @param environment [String] The environment of the file (\&quot;draft\&quot; or \&quot;published\&quot;).
         # @param path [String] The file system location of the file.
         # @param [Hash] opts the optional parameters
-        # @option opts [File] :file The file to upload.
+        # @option opts [File] :file 
         # @return [Array<(AssetFileMetadata, Integer, Hash)>] AssetFileMetadata data, response status code and response headers
         def create_with_http_info(environment, path, opts = {})
           if @api_client.config.debugging
@@ -182,7 +182,7 @@ module Hubspot
         # @param environment [String] The environment of the file (\&quot;draft\&quot; or \&quot;published\&quot;).
         # @param path [String] The file system location of the file.
         # @param [Hash] opts the optional parameters
-        # @option opts [File] :file The file to upload.
+        # @option opts [File] :file 
         # @return [AssetFileMetadata]
         def create_or_update(environment, path, opts = {})
           data, _status_code, _headers = create_or_update_with_http_info(environment, path, opts)
@@ -194,7 +194,7 @@ module Hubspot
         # @param environment [String] The environment of the file (\&quot;draft\&quot; or \&quot;published\&quot;).
         # @param path [String] The file system location of the file.
         # @param [Hash] opts the optional parameters
-        # @option opts [File] :file The file to upload.
+        # @option opts [File] :file 
         # @return [Array<(AssetFileMetadata, Integer, Hash)>] AssetFileMetadata data, response status code and response headers
         def create_or_update_with_http_info(environment, path, opts = {})
           if @api_client.config.debugging
