+++ lib/hubspot/codegen/cms/source_code/api/extract_api.rb
@@ -1,7 +1,7 @@
 =begin
 #CMS Source Code
 
-#Endpoints for interacting with files in the CMS Developer File System. These files include HTML templates, CSS, JS, modules, and other assets which are used to create CMS content.
+#No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 
 The version of the OpenAPI document: v3
 
@@ -21,36 +21,31 @@ module Hubspot
         def initialize(api_client = ApiClient.default)
           @api_client = api_client
         end
-        # Extracts a zip file
-        # Extracts a zip file in the file system. The zip file will be extracted in-place and not be deleted automatically.
-        # @param path [String] The file system location of the zip file.
+        # Extract a zip file
+        # Extract a zip file in the developer file system. Extraction status can be checked with the `/extract/async/tasks/taskId/status` endpoint below.
+        # @param file_extract_request [FileExtractRequest] 
         # @param [Hash] opts the optional parameters
-        # @return [nil]
-        def extract_by_path(path, opts = {})
-          extract_by_path_with_http_info(path, opts)
-          nil
+        # @return [TaskLocator]
+        def do_async(file_extract_request, opts = {})
+          data, _status_code, _headers = do_async_with_http_info(file_extract_request, opts)
+          data
         end
 
-        # Extracts a zip file
-        # Extracts a zip file in the file system. The zip file will be extracted in-place and not be deleted automatically.
-        # @param path [String] The file system location of the zip file.
+        # Extract a zip file
+        # Extract a zip file in the developer file system. Extraction status can be checked with the &#x60;/extract/async/tasks/taskId/status&#x60; endpoint below.
+        # @param file_extract_request [FileExtractRequest] 
         # @param [Hash] opts the optional parameters
-        # @return [Array<(nil, Integer, Hash)>] nil, response status code and response headers
-        def extract_by_path_with_http_info(path, opts = {})
+        # @return [Array<(TaskLocator, Integer, Hash)>] TaskLocator data, response status code and response headers
+        def do_async_with_http_info(file_extract_request, opts = {})
           if @api_client.config.debugging
-            @api_client.config.logger.debug 'Calling API: ExtractApi.extract_by_path ...'
+            @api_client.config.logger.debug 'Calling API: ExtractApi.do_async ...'
           end
-          # verify the required parameter 'path' is set
-          if @api_client.config.client_side_validation && path.nil?
-            fail ArgumentError, "Missing the required parameter 'path' when calling ExtractApi.extract_by_path"
+          # verify the required parameter 'file_extract_request' is set
+          if @api_client.config.client_side_validation && file_extract_request.nil?
+            fail ArgumentError, "Missing the required parameter 'file_extract_request' when calling ExtractApi.do_async"
           end
-          pattern = Regexp.new(/.+/)
-          if @api_client.config.client_side_validation && path !~ pattern
-            fail ArgumentError, "invalid value for 'path' when calling ExtractApi.extract_by_path, must conform to the pattern #{pattern}."
-          end
-
           # resource path
-          local_var_path = '/cms/v3/source-code/extract/{path}'.sub('{' + 'path' + '}', CGI.escape(path.to_s))
+          local_var_path = '/cms/v3/source-code/extract/async'
 
           # query parameters
           query_params = opts[:query_params] || {}
@@ -58,22 +53,27 @@ module Hubspot
           # header parameters
           header_params = opts[:header_params] || {}
           # HTTP header 'Accept' (if needed)
-          header_params['Accept'] = @api_client.select_header_accept(['*/*'])
+          header_params['Accept'] = @api_client.select_header_accept(['application/json', '*/*'])
+          # HTTP header 'Content-Type'
+          content_type = @api_client.select_header_content_type(['application/json'])
+          if !content_type.nil?
+              header_params['Content-Type'] = content_type
+          end
 
           # form parameters
           form_params = opts[:form_params] || {}
 
           # http body (model)
-          post_body = opts[:debug_body]
+          post_body = opts[:debug_body] || @api_client.object_to_http_body(file_extract_request)
 
           # return_type
-          return_type = opts[:debug_return_type]
+          return_type = opts[:debug_return_type] || 'TaskLocator'
 
           # auth_names
           auth_names = opts[:debug_auth_names] || ['oauth2']
 
           new_options = opts.merge(
-            :operation => :"ExtractApi.extract_by_path",
+            :operation => :"ExtractApi.do_async",
             :header_params => header_params,
             :query_params => query_params,
             :form_params => form_params,
@@ -84,7 +84,70 @@ module Hubspot
 
           data, status_code, headers = @api_client.call_api(:POST, local_var_path, new_options)
           if @api_client.config.debugging
-            @api_client.config.logger.debug "API called: ExtractApi#extract_by_path\nData: #{data.inspect}\nStatus code: #{status_code}\nHeaders: #{headers}"
+            @api_client.config.logger.debug "API called: ExtractApi#do_async\nData: #{data.inspect}\nStatus code: #{status_code}\nHeaders: #{headers}"
+          end
+          return data, status_code, headers
+        end
+
+        # Get extraction status
+        # Get the status of an extraction by the `taskId` returned from the initial `extract/async` request.
+        # @param task_id [Integer] The extraction task ID returned by the initial &#x60;extract/async&#x60; request.
+        # @param [Hash] opts the optional parameters
+        # @return [ActionResponse]
+        def get_async_status(task_id, opts = {})
+          data, _status_code, _headers = get_async_status_with_http_info(task_id, opts)
+          data
+        end
+
+        # Get extraction status
+        # Get the status of an extraction by the &#x60;taskId&#x60; returned from the initial &#x60;extract/async&#x60; request.
+        # @param task_id [Integer] The extraction task ID returned by the initial &#x60;extract/async&#x60; request.
+        # @param [Hash] opts the optional parameters
+        # @return [Array<(ActionResponse, Integer, Hash)>] ActionResponse data, response status code and response headers
+        def get_async_status_with_http_info(task_id, opts = {})
+          if @api_client.config.debugging
+            @api_client.config.logger.debug 'Calling API: ExtractApi.get_async_status ...'
+          end
+          # verify the required parameter 'task_id' is set
+          if @api_client.config.client_side_validation && task_id.nil?
+            fail ArgumentError, "Missing the required parameter 'task_id' when calling ExtractApi.get_async_status"
+          end
+          # resource path
+          local_var_path = '/cms/v3/source-code/extract/async/tasks/{taskId}/status'.sub('{' + 'taskId' + '}', CGI.escape(task_id.to_s))
+
+          # query parameters
+          query_params = opts[:query_params] || {}
+
+          # header parameters
+          header_params = opts[:header_params] || {}
+          # HTTP header 'Accept' (if needed)
+          header_params['Accept'] = @api_client.select_header_accept(['application/json', '*/*'])
+
+          # form parameters
+          form_params = opts[:form_params] || {}
+
+          # http body (model)
+          post_body = opts[:debug_body]
+
+          # return_type
+          return_type = opts[:debug_return_type] || 'ActionResponse'
+
+          # auth_names
+          auth_names = opts[:debug_auth_names] || ['oauth2']
+
+          new_options = opts.merge(
+            :operation => :"ExtractApi.get_async_status",
+            :header_params => header_params,
+            :query_params => query_params,
+            :form_params => form_params,
+            :body => post_body,
+            :auth_names => auth_names,
+            :return_type => return_type
+          )
+
+          data, status_code, headers = @api_client.call_api(:GET, local_var_path, new_options)
+          if @api_client.config.debugging
+            @api_client.config.logger.debug "API called: ExtractApi#get_async_status\nData: #{data.inspect}\nStatus code: #{status_code}\nHeaders: #{headers}"
           end
           return data, status_code, headers
         end
