+++ lib/hubspot/codegen/cms/source_code/api/validation_api.rb
@@ -1,7 +1,7 @@
 =begin
 #CMS Source Code
 
-#Endpoints for interacting with files in the CMS Developer File System. These files include HTML templates, CSS, JS, modules, and other assets which are used to create CMS content.
+#No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 
 The version of the OpenAPI document: v3
 
@@ -25,7 +25,7 @@ module Hubspot
         # Validates the file contents passed to the endpoint given a specified path and environment. Accepts multipart/form-data content type.
         # @param path [String] The file system location of the file.
         # @param [Hash] opts the optional parameters
-        # @option opts [File] :file The file to validate.
+        # @option opts [File] :file 
         # @return [Error]
         def do_validate(path, opts = {})
           data, _status_code, _headers = do_validate_with_http_info(path, opts)
@@ -36,7 +36,7 @@ module Hubspot
         # Validates the file contents passed to the endpoint given a specified path and environment. Accepts multipart/form-data content type.
         # @param path [String] The file system location of the file.
         # @param [Hash] opts the optional parameters
-        # @option opts [File] :file The file to validate.
+        # @option opts [File] :file 
         # @return [Array<(Error, Integer, Hash)>] Error data, response status code and response headers
         def do_validate_with_http_info(path, opts = {})
           if @api_client.config.debugging
