+++ lib/hubspot/codegen/cms/source_code/models/task_locator.rb
@@ -1,7 +1,7 @@
 =begin
 #CMS Source Code
 
-#Endpoints for interacting with files in the CMS Developer File System. These files include HTML templates, CSS, JS, modules, and other assets which are used to create CMS content.
+#No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 
 The version of the OpenAPI document: v3
